requirements.txt
# Data Collection
requests==2.31.0
beautifulsoup4==4.12.0
pandas==2.0.0
numpy==1.24.3
selenium==4.15.2
webdriver-manager==4.0.1

# Database
py2neo==2021.2.3
neo4j==5.14.0

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1

# Development Tools
python-dotenv==1.0.0
pyyaml==6.0.1
jupyter==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.10.1
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0